PRO-1234: Test
