+++ src/componentTest/java/uk/gov/hmcts/probate/component/clients/v2/ccd/CcdClientApiTest.java
@@ -1,0 +1,375 @@
+package uk.gov.hmcts.probate.component.clients.v2.ccd;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.google.common.collect.ImmutableMap;
+import org.assertj.core.util.Lists;
+import org.junit.Before;
+import org.junit.Rule;
+import org.junit.Test;
+import org.junit.rules.ExpectedException;
+import org.junit.runner.RunWith;
+import org.mockito.ArgumentCaptor;
+import org.mockito.Captor;
+import org.mockito.Mock;
+import org.mockito.junit.MockitoJUnitRunner;
+import uk.gov.hmcts.probate.security.SecurityDTO;
+import uk.gov.hmcts.probate.services.submit.clients.v2.ccd.CaseDetailsToCaseDataMapper;
+import uk.gov.hmcts.probate.services.submit.clients.v2.ccd.CcdClientApi;
+import uk.gov.hmcts.probate.services.submit.clients.v2.ccd.CcdElasticSearchQueryBuilder;
+import uk.gov.hmcts.probate.services.submit.core.SearchFieldFactory;
+import uk.gov.hmcts.reform.ccd.client.CaseAccessApi;
+import uk.gov.hmcts.reform.ccd.client.CoreCaseDataApi;
+import uk.gov.hmcts.reform.ccd.client.model.*;
+import uk.gov.hmcts.reform.probate.model.cases.*;
+import uk.gov.hmcts.reform.probate.model.cases.grantofrepresentation.GrantOfRepresentationData;
+import uk.gov.hmcts.reform.probate.model.cases.grantofrepresentation.GrantType;
+
+import java.time.LocalDateTime;
+import java.util.List;
+import java.util.Optional;
+
+import static org.hamcrest.Matchers.is;
+import static org.hamcrest.Matchers.notNullValue;
+import static org.junit.Assert.assertThat;
+import static org.mockito.ArgumentMatchers.eq;
+import static org.mockito.Mockito.*;
+import static uk.gov.hmcts.reform.probate.model.cases.CaseType.GRANT_OF_REPRESENTATION;
+import static uk.gov.hmcts.reform.probate.model.cases.JurisdictionId.PROBATE;
+
+@RunWith(MockitoJUnitRunner.class)
+public class CcdClientApiTest {
+
+    private static final Long CASE_ID = 123456789L;
+
+    private static final CaseState STATE = CaseState.DRAFT;
+
+    private static final String AUTHORIZATION = "AUTHXXXXXXXXXXXXX";
+
+    private static final String SERVICE_AUTHORIZATION = "SERVAUTHZZZZZZZZZZZZZ";
+
+    private static final String USER_ID = "33";
+
+    private static final String TOKEN = "TOKENXXXX123466LLDJH";
+
+    private static final String APPLICANT_EMAIL = "test@test.com";
+
+    private static final String EMAIL_QUERY_PARAM = "case.primaryApplicantEmailAddress";
+
+    private static final EventId CREATE_DRAFT = EventId.GOP_CREATE_DRAFT;
+
+    private static final EventId UPDATE_DRAFT = EventId.GOP_UPDATE_DRAFT;
+    public static final String INVITATION_ID = "12334345";
+    public static final String INVIATION_QUERY = "inviationQuery";
+
+    @Mock
+    private CoreCaseDataApi mockCoreCaseDataApi;
+    @Mock
+    private CaseAccessApi mockCaseAccessApi;
+
+    @Mock
+    private SearchFieldFactory searchFieldFactory;
+
+    @Mock
+    private CcdElasticSearchQueryBuilder mockInvitationElasticSearchQueryBuilder;
+
+    private CcdClientApi ccdClientApi;
+
+    @Rule
+    public ExpectedException exception = ExpectedException.none();
+
+    private SecurityDTO securityDTO;
+
+    private CaseData caseData;
+
+    private StartEventResponse startEventResponse;
+
+    private CaseDetails caseDetails;
+
+    private CaseDataContent caseDataContent;
+
+    private String inviteField = "inviteField";
+
+    @Captor
+    private ArgumentCaptor<UserId> userIdCaptor;
+
+    @Before
+    public void setUp() {
+        ccdClientApi = new CcdClientApi(mockCoreCaseDataApi, mockCaseAccessApi, new CaseDetailsToCaseDataMapper(new ObjectMapper()), searchFieldFactory, mockInvitationElasticSearchQueryBuilder);
+
+        when(searchFieldFactory.getEsSearchFieldName(CaseType.GRANT_OF_REPRESENTATION)).thenReturn("primaryApplicantEmailAddress");
+
+        when(searchFieldFactory.getSearchApplicantEmailFieldName()).thenReturn("primaryApplicantEmailAddress");
+
+
+        String inviteField = "inviteField";
+
+        when(searchFieldFactory.getSearchInviteFieldName()).thenReturn(inviteField);
+
+        securityDTO = SecurityDTO.builder()
+                .authorisation(AUTHORIZATION)
+                .serviceAuthorisation(SERVICE_AUTHORIZATION)
+                .userId(USER_ID)
+                .build();
+
+        caseData = new GrantOfRepresentationData();
+        startEventResponse = StartEventResponse.builder()
+                .token(TOKEN)
+                .build();
+
+        caseDetails = CaseDetails.builder()
+                .id(CASE_ID)
+                .state(STATE.getName())
+                .caseTypeId(GRANT_OF_REPRESENTATION.getName())
+                .createdDate(LocalDateTime.now())
+                .data(ImmutableMap.of("applicationType", ApplicationType.PERSONAL,
+                        "caseType", GrantType.INTESTACY))
+                .build();
+
+        caseDataContent = CaseDataContent.builder()
+                .eventToken(TOKEN)
+                .event(Event.builder()
+                        .id(CREATE_DRAFT.getName())
+                        .description("Probate application")
+                        .summary("Probate application")
+                        .build())
+                .data(caseData)
+                .build();
+    }
+
+    @Test
+    public void shouldCreateCase() {
+        when(mockCoreCaseDataApi.startForCitizen(AUTHORIZATION, SERVICE_AUTHORIZATION, USER_ID, PROBATE.name(),
+                GRANT_OF_REPRESENTATION.getName(), CREATE_DRAFT.getName())).thenReturn(startEventResponse);
+
+        when(mockCoreCaseDataApi.submitForCitizen(eq(AUTHORIZATION), eq(SERVICE_AUTHORIZATION), eq(USER_ID), eq(PROBATE.name()),
+                eq(GRANT_OF_REPRESENTATION.getName()), eq(false), eq(caseDataContent))).thenReturn(caseDetails);
+
+        ProbateCaseDetails caseResponse = ccdClientApi.createCase(caseData, CREATE_DRAFT, securityDTO);
+
+        assertThat(caseResponse, is(notNullValue()));
+        assertThat(caseResponse.getCaseInfo().getCaseId(), is(CASE_ID.toString()));
+        assertThat(caseResponse.getCaseInfo().getState(), is(STATE));
+        verify(mockCoreCaseDataApi, times(1)).startForCitizen(AUTHORIZATION, SERVICE_AUTHORIZATION, USER_ID, PROBATE.name(),
+                GRANT_OF_REPRESENTATION.getName(), CREATE_DRAFT.getName());
+        verify(mockCoreCaseDataApi, times(1)).submitForCitizen(eq(AUTHORIZATION), eq(SERVICE_AUTHORIZATION), eq(USER_ID), eq(PROBATE.name()),
+                eq(GRANT_OF_REPRESENTATION.getName()), eq(false), eq(caseDataContent));
+    }
+
+
+    @Test
+    public void shouldCreateCaseCaseAsCaseworker() {
+        when(mockCoreCaseDataApi.startForCaseworker(AUTHORIZATION, SERVICE_AUTHORIZATION, USER_ID, PROBATE.name(),
+                GRANT_OF_REPRESENTATION.getName(), CREATE_DRAFT.getName())).thenReturn(startEventResponse);
+
+        when(mockCoreCaseDataApi.submitForCaseworker(eq(AUTHORIZATION), eq(SERVICE_AUTHORIZATION), eq(USER_ID), eq(PROBATE.name()),
+                eq(GRANT_OF_REPRESENTATION.getName()), eq(false), eq(caseDataContent))).thenReturn(caseDetails);
+
+        ProbateCaseDetails caseResponse = ccdClientApi.createCaseAsCaseworker(
+                caseData, CREATE_DRAFT, securityDTO);
+
+        assertThat(caseResponse, is(notNullValue()));
+        assertThat(caseResponse.getCaseInfo().getCaseId(), is(CASE_ID.toString()));
+        assertThat(caseResponse.getCaseInfo().getState(), is(STATE));
+        verify(mockCoreCaseDataApi, times(1)).startForCaseworker(AUTHORIZATION, SERVICE_AUTHORIZATION, USER_ID, PROBATE.name(),
+                GRANT_OF_REPRESENTATION.getName(), CREATE_DRAFT.getName());
+        verify(mockCoreCaseDataApi, times(1)).submitForCaseworker(eq(AUTHORIZATION), eq(SERVICE_AUTHORIZATION), eq(USER_ID), eq(PROBATE.name()),
+                eq(GRANT_OF_REPRESENTATION.getName()), eq(false), eq(caseDataContent));
+    }
+
+    @Test
+    public void shouldUpdateCase() {
+        caseDataContent.getEvent().setId(UPDATE_DRAFT.getName());
+
+        when(mockCoreCaseDataApi.startEventForCitizen(AUTHORIZATION, SERVICE_AUTHORIZATION, USER_ID, PROBATE.name(),
+                GRANT_OF_REPRESENTATION.getName(), CASE_ID.toString(), UPDATE_DRAFT.getName())).thenReturn(startEventResponse);
+
+        when(mockCoreCaseDataApi.submitEventForCitizen(eq(AUTHORIZATION), eq(SERVICE_AUTHORIZATION), eq(USER_ID), eq(PROBATE.name()),
+                eq(GRANT_OF_REPRESENTATION.getName()), eq(CASE_ID.toString()), eq(false), eq(caseDataContent))).thenReturn(caseDetails);
+
+        ProbateCaseDetails caseResponse = ccdClientApi.updateCase(CASE_ID.toString(), caseData, UPDATE_DRAFT, securityDTO);
+
+        assertThat(caseResponse, is(notNullValue()));
+        assertThat(caseResponse.getCaseInfo().getCaseId(), is(CASE_ID.toString()));
+        assertThat(caseResponse.getCaseInfo().getState(), is(STATE));
+        verify(mockCoreCaseDataApi, times(1)).startEventForCitizen(AUTHORIZATION, SERVICE_AUTHORIZATION, USER_ID, PROBATE.name(),
+                GRANT_OF_REPRESENTATION.getName(), CASE_ID.toString(), UPDATE_DRAFT.getName());
+        verify(mockCoreCaseDataApi, times(1)).submitEventForCitizen(eq(AUTHORIZATION), eq(SERVICE_AUTHORIZATION), eq(USER_ID), eq(PROBATE.name()),
+                eq(GRANT_OF_REPRESENTATION.getName()), eq(CASE_ID.toString()), eq(false), eq(caseDataContent));
+    }
+
+    @Test
+    public void shouldUpdateCaseAsCaseWorker() {
+        caseDataContent.getEvent().setId(UPDATE_DRAFT.getName());
+
+        when(mockCoreCaseDataApi.startEventForCaseWorker(AUTHORIZATION, SERVICE_AUTHORIZATION, USER_ID, PROBATE.name(),
+                GRANT_OF_REPRESENTATION.getName(), CASE_ID.toString(), UPDATE_DRAFT.getName())).thenReturn(startEventResponse);
+
+        when(mockCoreCaseDataApi.submitEventForCaseWorker(eq(AUTHORIZATION), eq(SERVICE_AUTHORIZATION), eq(USER_ID), eq(PROBATE.name()),
+                eq(GRANT_OF_REPRESENTATION.getName()), eq(CASE_ID.toString()), eq(false), eq(caseDataContent))).thenReturn(caseDetails);
+
+        ProbateCaseDetails caseResponse = ccdClientApi.updateCaseAsCaseworker(CASE_ID.toString(), caseData, UPDATE_DRAFT, securityDTO);
+
+        assertThat(caseResponse, is(notNullValue()));
+        assertThat(caseResponse.getCaseInfo().getCaseId(), is(CASE_ID.toString()));
+        assertThat(caseResponse.getCaseInfo().getState(), is(STATE));
+        verify(mockCoreCaseDataApi, times(1)).startEventForCaseWorker(AUTHORIZATION, SERVICE_AUTHORIZATION, USER_ID, PROBATE.name(),
+                GRANT_OF_REPRESENTATION.getName(), CASE_ID.toString(), UPDATE_DRAFT.getName());
+        verify(mockCoreCaseDataApi, times(1)).submitEventForCaseWorker(eq(AUTHORIZATION), eq(SERVICE_AUTHORIZATION), eq(USER_ID), eq(PROBATE.name()),
+                eq(GRANT_OF_REPRESENTATION.getName()), eq(CASE_ID.toString()), eq(false), eq(caseDataContent));
+    }
+
+    @Test
+    public void shouldFindCase() {
+
+        String queryString = "queryString";
+        when(mockInvitationElasticSearchQueryBuilder.buildQuery(APPLICANT_EMAIL, "primaryApplicantEmailAddress")).thenReturn(queryString);
+
+        when(mockCoreCaseDataApi.searchCases(eq(AUTHORIZATION), eq(SERVICE_AUTHORIZATION),
+                eq(GRANT_OF_REPRESENTATION.getName()), eq(queryString))).thenReturn(SearchResult.builder().cases(Lists.newArrayList(caseDetails)).build());
+
+
+        Optional<ProbateCaseDetails> optionalCaseResponse = ccdClientApi.findCase(APPLICANT_EMAIL, GRANT_OF_REPRESENTATION, securityDTO);
+
+        ProbateCaseDetails caseResponse = optionalCaseResponse.get();
+        assertThat(caseResponse, is(notNullValue()));
+        assertThat(caseResponse.getCaseInfo().getCaseId(), is(CASE_ID.toString()));
+        assertThat(caseResponse.getCaseInfo().getState(), is(STATE));
+        verify(mockCoreCaseDataApi, times(1)).searchCases(eq(AUTHORIZATION), eq(SERVICE_AUTHORIZATION),
+                eq(GRANT_OF_REPRESENTATION.getName()), eq(queryString));
+    }
+
+    @Test
+    public void shouldFindCaseByApplicantEmail() {
+
+        String queryString = "queryString";
+        when(mockInvitationElasticSearchQueryBuilder.buildQuery(APPLICANT_EMAIL, "primaryApplicantEmailAddress")).thenReturn(queryString);
+
+        when(mockCoreCaseDataApi.searchCases(eq(AUTHORIZATION), eq(SERVICE_AUTHORIZATION),
+                eq(GRANT_OF_REPRESENTATION.getName()), eq(queryString))).thenReturn(SearchResult.builder().cases(Lists.newArrayList(caseDetails)).build());
+
+
+        Optional<ProbateCaseDetails> optionalCaseResponse = ccdClientApi.findCaseByApplicantEmail(APPLICANT_EMAIL, GRANT_OF_REPRESENTATION, securityDTO);
+
+        ProbateCaseDetails caseResponse = optionalCaseResponse.get();
+        assertThat(caseResponse, is(notNullValue()));
+        assertThat(caseResponse.getCaseInfo().getCaseId(), is(CASE_ID.toString()));
+        assertThat(caseResponse.getCaseInfo().getState(), is(STATE));
+        verify(mockCoreCaseDataApi, times(1)).searchCases(eq(AUTHORIZATION), eq(SERVICE_AUTHORIZATION),
+                eq(GRANT_OF_REPRESENTATION.getName()), eq(queryString));
+    }
+
+    @Test
+    public void shouldFindAllCase() {
+
+        String queryString = "queryString";
+        when(mockInvitationElasticSearchQueryBuilder.buildFindAllCasesQuery()).thenReturn(queryString);
+
+        when(mockCoreCaseDataApi.searchCases(eq(AUTHORIZATION), eq(SERVICE_AUTHORIZATION),
+                eq(GRANT_OF_REPRESENTATION.getName()), eq(queryString))).thenReturn(SearchResult.builder().cases(Lists.newArrayList(caseDetails)).build());
+
+        List<ProbateCaseDetails> results = ccdClientApi.findCases(GRANT_OF_REPRESENTATION, securityDTO);
+
+        CaseInfo caseInfo = results.stream().findFirst().get().getCaseInfo();
+        assertThat(results, is(notNullValue()));
+        assertThat(caseInfo.getCaseId(), is(CASE_ID.toString()));
+        assertThat(caseInfo.getState(), is(STATE));
+        verify(mockCoreCaseDataApi, times(1)).searchCases(eq(AUTHORIZATION), eq(SERVICE_AUTHORIZATION),
+                eq(GRANT_OF_REPRESENTATION.getName()), eq(queryString));
+    }
+
+    @Test
+    public void shouldFindCaseById() {
+        when(mockCoreCaseDataApi.getCase(eq(AUTHORIZATION), eq(SERVICE_AUTHORIZATION), eq(CASE_ID.toString())))
+                .thenReturn(caseDetails);
+
+        Optional<ProbateCaseDetails> optionalCaseResponse = ccdClientApi.findCaseById(CASE_ID.toString(), securityDTO);
+
+        ProbateCaseDetails caseResponse = optionalCaseResponse.get();
+        assertThat(caseResponse, is(notNullValue()));
+        assertThat(caseResponse.getCaseInfo().getCaseId(), is(CASE_ID.toString()));
+        assertThat(caseResponse.getCaseInfo().getState(), is(STATE));
+        verify(mockCoreCaseDataApi, times(1)).getCase(eq(AUTHORIZATION), eq(SERVICE_AUTHORIZATION), eq(CASE_ID.toString()));
+    }
+
+    @Test
+    public void shouldReturnEmptyOptionalWhenCaseNotFoundById() {
+        when(mockCoreCaseDataApi.getCase(eq(AUTHORIZATION), eq(SERVICE_AUTHORIZATION), eq(CASE_ID.toString())))
+                .thenReturn(null);
+
+        Optional<ProbateCaseDetails> optionalCaseResponse = ccdClientApi.findCaseById(CASE_ID.toString(), securityDTO);
+
+        assertThat(optionalCaseResponse.isPresent(), is(false));
+        verify(mockCoreCaseDataApi, times(1)).getCase(eq(AUTHORIZATION), eq(SERVICE_AUTHORIZATION), eq(CASE_ID.toString()));
+    }
+
+    @Test
+    public void shouldFindCaseByInvitationId() {
+
+
+        when(mockInvitationElasticSearchQueryBuilder.buildQuery(INVITATION_ID, inviteField)).thenReturn(INVIATION_QUERY);
+
+        when(mockCoreCaseDataApi.searchCases(eq(AUTHORIZATION), eq(SERVICE_AUTHORIZATION), eq(GRANT_OF_REPRESENTATION.getName()), eq(INVIATION_QUERY))).thenReturn(SearchResult.builder().cases(Lists.newArrayList(caseDetails)).build());
+
+        Optional<ProbateCaseDetails> optionalCaseResponse = ccdClientApi.findCaseByInviteId(INVITATION_ID, GRANT_OF_REPRESENTATION, securityDTO);
+
+        ProbateCaseDetails caseResponse = optionalCaseResponse.get();
+        assertThat(caseResponse, is(notNullValue()));
+        assertThat(caseResponse.getCaseInfo().getCaseId(), is(CASE_ID.toString()));
+        assertThat(caseResponse.getCaseInfo().getState(), is(STATE));
+        verify(mockCoreCaseDataApi, times(1)).searchCases(eq(AUTHORIZATION), eq(SERVICE_AUTHORIZATION), eq(GRANT_OF_REPRESENTATION.getName()), eq(INVIATION_QUERY));
+    }
+
+
+    @Test
+    public void shouldReturnEmptyOptionalWhenReturningNullOnSearch() {
+        String queryString = "queryString";
+        when(mockInvitationElasticSearchQueryBuilder.buildQuery(APPLICANT_EMAIL, "primaryApplicantEmailAddress")).thenReturn(queryString);
+
+        when(mockCoreCaseDataApi.searchCases(eq(AUTHORIZATION), eq(SERVICE_AUTHORIZATION),
+                eq(GRANT_OF_REPRESENTATION.getName()), eq(queryString))).thenReturn(SearchResult.builder().cases(null).build());
+
+
+        Optional<ProbateCaseDetails> optionalCaseResponse = ccdClientApi.findCase(APPLICANT_EMAIL, GRANT_OF_REPRESENTATION, securityDTO);
+
+
+        assertThat(optionalCaseResponse.isPresent(), is(false));
+    }
+
+    @Test
+    public void shouldReturnEmptyOptionalWhenReturningNullOnSearchInvitation() {
+        String invitationId = INVITATION_ID;
+        String invitationQuery =
+                INVIATION_QUERY;
+        when(mockInvitationElasticSearchQueryBuilder.buildQuery(invitationId, inviteField)).thenReturn(invitationQuery);
+
+        when(mockCoreCaseDataApi.searchCases(eq(AUTHORIZATION), eq(SERVICE_AUTHORIZATION), eq(GRANT_OF_REPRESENTATION.getName()), eq(invitationQuery))).thenReturn(SearchResult.builder().cases(null).build());
+
+        Optional<ProbateCaseDetails> optionalCaseResponse = ccdClientApi.findCaseByInviteId(invitationId, GRANT_OF_REPRESENTATION, securityDTO);
+
+        assertThat(optionalCaseResponse.isPresent(), is(false));
+    }
+
+    @Test
+    public void shouldThrowIllegalStateExceptionWhenFindingMoreThanOneCase() {
+        exception.expect(IllegalStateException.class);
+        exception.expectMessage("Multiple cases exist with applicant email provided!");
+
+        String queryString = "queryString";
+        when(mockInvitationElasticSearchQueryBuilder.buildQuery(APPLICANT_EMAIL, "primaryApplicantEmailAddress")).thenReturn(queryString);
+
+        when(mockCoreCaseDataApi.searchCases(eq(AUTHORIZATION), eq(SERVICE_AUTHORIZATION),
+                eq(GRANT_OF_REPRESENTATION.getName()), eq(queryString))).thenReturn(SearchResult.builder().cases(Lists.newArrayList(caseDetails, caseDetails)).build());
+
+
+        ccdClientApi.findCase(APPLICANT_EMAIL, GRANT_OF_REPRESENTATION, securityDTO);
+    }
+
+    @Test
+    public void shouldGrantAccessToCaseByApplicantEmail() {
+
+       ccdClientApi.grantAccessForCase(CaseType.GRANT_OF_REPRESENTATION, CASE_ID.toString(), APPLICANT_EMAIL, securityDTO);
+
+        verify(mockCaseAccessApi, times(1)).grantAccessToCase(eq(AUTHORIZATION), eq(SERVICE_AUTHORIZATION),         eq(USER_ID), eq(PROBATE.name()), eq(GRANT_OF_REPRESENTATION.getName()), eq(CASE_ID.toString()), userIdCaptor.capture());
+        assertThat(userIdCaptor.getValue().getId(), is(APPLICANT_EMAIL));
+    }
+}
