+++ src/componentTest/resources/application.properties
@@ -1,0 +1,1 @@
+probate.submit.url=${TEST_URL:http://localhost:8181}
