+++ src/componentTest/resources/files/v2/intestacyGrantOfRepresentationEmpty.json
@@ -1,0 +1,3 @@
+{
+  "type": "GrantOfRepresentation"
+}
