+++ src/componentTest/resources/files/v2/intestacyGrantOfRepresentationInvalid.json
@@ -1,0 +1,65 @@
+{
+  "type": "GrantOfRepresentation",
+  "applicationType": "Personal",
+  "primaryApplicantEmailAddress": "jon.snow@thenorth.com",
+  "primaryApplicantForenames": "Jon",
+  "primaryApplicantSurname": "Snow",
+  "primaryApplicantAddress": {
+    "AddressLine1": "Pret a Manger St. Georges Hospital Blackshaw Road London SW17 0QT"
+  },
+  "primaryApplicantFreeTextAddress": "Pret a Manger St. Georges Hospital Blackshaw Road",
+  "primaryApplicantAddressFound": "Yes",
+  "primaryApplicantPhoneNumber": "123455678",
+  "primaryApplicantRelationshipToDeceased": "adoptedChild",
+  "primaryApplicantAdoptionInEnglandOrWales": "Yes",
+  "deceasedForenames": "N",
+  "deceasedSurname": "Stark",
+  "deceasedDateOfBirth": "1954-01-01",
+  "deceasedDateOfDeath": "2018-01-01",
+  "deceasedAddress": {
+    "AddressLine1": "Winterfell, Westeros"
+  },
+  "deceasedFreeTextAddress": "Winterfell, Westeros",
+  "deceasedAddressFound": "Yes",
+  "deceasedAnyOtherNames": "No",
+  "deceasedAliasNameList": [
+    {
+      "value": {
+        "Forenames": "King",
+        "LastName": "North"
+      }
+    }
+  ],
+  "deceasedMartialStatus": "marriedCivilPartnership",
+  "deceasedDivorcedInEnglandOrWales": "No",
+  "deceasedSpouseNotApplyingReason": "mentallyIncapable",
+  "deceasedOtherChildren": "Yes",
+  "childrenOverEighteenSurvived": "Yes",
+  "childrenDied": "No",
+  "grandChildrenSurvivedUnderEighteen": "No",
+  "deceasedAnyChildren": "No",
+  "registryLocation": "Birmingham",
+  "assetsOverseasNetValue": "10050",
+  "deceasedHasAssetsOutsideUK": "Yes",
+  "ihtFormId": "IHT205",
+  "ihtFormCompletedOnline": "Yes",
+  "ihtGrossValue": "100000",
+  "ihtNetValue": "100000",
+  "ihtReferenceNumber": "GOT123456",
+  "extraCopiesOfGrant": 5,
+  "outsideUKGrantCopies": 6,
+  "payments": [
+    {
+      "value": {
+        "status": "Success",
+        "date": "2018-12-03",
+        "reference": "RC-1537-1988-5489-1985",
+        "amount": "22050",
+        "method": "online",
+        "transactionId": "r23k178busa0rp2mh27m0vchja",
+        "siteId": "P223"
+      }
+    }
+  ],
+  "uploadDocumentUrl": "http://document-management/document/12345"
+}
