+++ src/componentTest/resources/files/v2/intestacyGrantOfRepresentation_caseDetails.json
@@ -1,0 +1,73 @@
+{
+	  "caseData": {
+	    "type": "GrantOfRepresentation",
+	    "caseType": "intestacy",
+	    "applicationType": "Personal",
+	    "primaryApplicantEmailAddress": "jsnow@bbc.co.uk",
+	    "primaryApplicantForenames": "Jon",
+	    "primaryApplicantSurname": "Snow",
+	    "primaryApplicantAddress": {
+	      "AddressLine1": "Pret a Manger St. Georges Hospital Blackshaw Road London SW17 0QT"
+	    },
+	    "primaryApplicantFreeTextAddress": "Pret a Manger St. Georges Hospital Blackshaw Road",
+	    "primaryApplicantAddressFound": "Yes",
+	    "primaryApplicantPhoneNumber": "123455678",
+	    "primaryApplicantRelationshipToDeceased": "adoptedChild",
+	    "primaryApplicantAdoptionInEnglandOrWales": "Yes",
+	    "payments": [
+	        {
+	          "value": {
+	            "status": "Success",
+	            "date": "2018-12-03",
+	            "reference": "RC-1537-1988-5489-1985",
+	            "amount": "22050",
+	            "method": "online",
+	            "transactionId": "r23k178busa0rp2mh27m0vchja",
+	            "siteId": "P223"
+	          }
+	        }
+	      ],
+	    "deceasedForenames": "Ned",
+	    "deceasedSurname": "Stark",
+	    "deceasedDateOfBirth": "1930-01-01",
+	    "deceasedDateOfDeath": "2018-01-01",
+	    "deceasedAddress": {
+	      "AddressLine1": "Winterfell, Westeros"
+	    },
+	    "deceasedFreeTextAddress": "Winterfell, Westeros",
+	    "deceasedAddressFound": "Yes",
+	    "deceasedAnyOtherNames": "No",
+	    "deceasedAliasNameList": [
+	      {
+	        "value": {
+	          "Forenames": "King",
+	          "LastName": "North"
+	        }
+	      }
+	    ],
+	    "deceasedMartialStatus": "marriedCivilPartnership",
+	    "deceasedDivorcedInEnglandOrWales": "No",
+	    "deceasedSpouseNotApplyingReason": "mentallyIncapable",
+	    "deceasedOtherChildren": "Yes",
+	    "childrenOverEighteenSurvived": "Yes",
+	    "childrenDied": "No",
+	    "grandChildrenSurvivedUnderEighteen": "No",
+	    "deceasedAnyChildren": "No",
+	    "registryLocation": "Birmingham",
+	    "assetsOverseasNetValue": "10050",
+	    "deceasedHasAssetsOutsideUK": "Yes",
+	    "ihtFormId": "IHT205",
+	    "ihtFormCompletedOnline": "Yes",
+	    "ihtGrossValue": "100000",
+	    "ihtNetValue": "100000",
+	    "ihtReferenceNumber": "GOT123456",
+	    "declarationCheckbox": "Yes",
+	    "extraCopiesOfGrant": 5,
+	    "outsideUKGrantCopies": 6,
+	    "uploadDocumentUrl": "http://document-management/document/12345"
+	  },
+	  "caseInfo": {
+	    "caseId": "12323213323",
+	    "state": "PAAppCreated"
+	  }
+	}
