+++ src/componentTest/resources/uk/gov/hmcts/probate/services/submit/git-test.properties
@@ -1,0 +1,2 @@
+git.commit.id=0773f129ad51c4a23a49fec96fec0888883443f6
+git.commit.time=2018-05-23T13:59+1234
